Use disposable scopes for test state restoration
