+++ Tests/Runtime/TestUtilities/CleanupScope.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: aa1ba17eb1d9400fb6ff4696715fef2c
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData:
+  assetBundleName:
+  assetBundleVariant:
